Clean up torque modules and created control loop for CMG
