+++ space_station_gnc/CMakeLists.txt
@@ -29,43 +29,39 @@ include_directories(${OpenCV_INCLUDE_DIRS} ${SDL2_INCLUDE_DIRS})
 add_executable(space_station_gnc_node src/space_station_gnc.cpp)
 add_executable(demo1a_nauka_incident_estimate src/demo1a_nauka_incident_estimate.cpp)
 add_executable(demo1b_crisis_mainengine src/demo1b_crisis_mainengine.cpp)
+add_executable(demo1c_small_incident src/demo1c_small_incident.cpp)
 add_executable(torque_collector src/torque_collector.cpp)
 add_executable(physics_motion src/physics_motion.cpp)
 add_executable(physics_sensor src/physics_sensor.cpp)
 add_executable(sense_estimate src/sense_estimate.cpp)
 add_executable(control_torque src/control_torque.cpp)
-add_executable(control_torque_act src/control_torque_act.cpp)
 ament_target_dependencies(space_station_gnc_node rclcpp std_msgs)
 ament_target_dependencies(demo1a_nauka_incident_estimate rclcpp std_msgs tf2_geometry_msgs tf2 tf2_ros OpenCV)
 ament_target_dependencies(demo1b_crisis_mainengine rclcpp std_msgs tf2_geometry_msgs tf2 tf2_ros OpenCV)
+ament_target_dependencies(demo1c_small_incident rclcpp std_msgs tf2_geometry_msgs tf2 tf2_ros OpenCV)
 ament_target_dependencies(torque_collector rclcpp std_msgs tf2 tf2_ros tf2_geometry_msgs)
 ament_target_dependencies(physics_motion rclcpp std_msgs tf2 tf2_ros tf2_geometry_msgs visualization_msgs)
 ament_target_dependencies(physics_sensor rclcpp std_msgs tf2 tf2_ros tf2_geometry_msgs visualization_msgs)
 ament_target_dependencies(sense_estimate rclcpp std_msgs tf2 tf2_ros nav_msgs tf2_geometry_msgs visualization_msgs)
 ament_target_dependencies(control_torque rclcpp std_msgs tf2 tf2_ros nav_msgs tf2_geometry_msgs visualization_msgs)
-ament_target_dependencies(control_torque_act rclcpp std_msgs tf2 tf2_ros nav_msgs tf2_geometry_msgs visualization_msgs)
 
 # Link OpenCV libraries
 target_link_libraries(demo1a_nauka_incident_estimate ${OpenCV_LIBRARIES} ${SDL2_LIBRARIES})
 target_link_libraries(demo1b_crisis_mainengine ${OpenCV_LIBRARIES} ${SDL2_LIBRARIES})
+target_link_libraries(demo1c_small_incident ${OpenCV_LIBRARIES} ${SDL2_LIBRARIES})
 
 # source by c++
 install(TARGETS
   space_station_gnc_node
   demo1a_nauka_incident_estimate
   demo1b_crisis_mainengine
+  demo1c_small_incident
   torque_collector
   physics_motion
   physics_sensor
   sense_estimate
   control_torque
-  control_torque_act
   DESTINATION lib/${PROJECT_NAME})
-
-# source by python
-#install(PROGRAMS
-#  src/space_station_gnc/demo_crisis.py
-#  DESTINATION lib/${PROJECT_NAME})
 
 #launch files
 install(
@@ -78,9 +74,6 @@ install(DIRECTORY urdf/
   DESTINATION share/${PROJECT_NAME}/urdf
 )
 
-#install(DIRECTORY meshes/
-#  DESTINATION share/${PROJECT_NAME}/meshes
-#)
 install(DIRECTORY urdf/meshes/
   DESTINATION share/${PROJECT_NAME}/urdf/meshes
 )
